git-svn-id: http://svn.leocad.org/tags/leocad-0.76@1059 c7d43263-9d01-0410-8a33-9dba5d9f93d6

+++ win/Leocad.cpp
@@ -1,7 +1,7 @@
 // LeoCAD.cpp : Defines the class behaviors for the application.
 //
 
-#include "lc_global.h"
+#include "stdafx.h"
 #include "LeoCAD.h"
 
 #include "MainFrm.h"
@@ -10,12 +10,14 @@
 #include <wininet.h>
 #include <process.h>
 #include "project.h"
+#include "globals.h"
 #include "system.h"
+#include "pieceinf.h" // TODO: remove
+#include "config.h"
 #include "mainwnd.h"
 #include "library.h"
 #include "keyboard.h"
 #include "lc_application.h"
-#include "preview.h"
 
 #ifdef _DEBUG
 #define new DEBUG_NEW
@@ -65,17 +67,16 @@ static void CheckForUpdates(void* Data)
 	if (hHttpFile)
 	{
 		if(HttpQueryInfo(hHttpFile,HTTP_QUERY_CONTENT_LENGTH, szSizeBuffer, &dwLengthSizeBuffer, NULL))
-		{
+		{	 
 			dwFileSize = atol(szSizeBuffer);
-			LPSTR szContents = Contents.GetBuffer(dwFileSize+1);
-			szContents[dwFileSize] = 0;
-
+			LPSTR szContents = Contents.GetBuffer(dwFileSize);
+			
 			if (InternetReadFile(hHttpFile, szContents, dwFileSize, &dwBytesRead))
 			{
 				float ver;
 				int lib;
 
-				if (sscanf(szContents, "%f %d", &ver, &lib) == 2)
+				if (sscanf (szContents, "%f %d", &ver, &lib) == 2)
 				{
 					CString str;
 					bool Update = false;
@@ -100,11 +101,11 @@ static void CheckForUpdates(void* Data)
 					{
 						if (Update)
 						{
-							str += "Would you like to visit the LeoCAD downloads page now?\n";
+							str += "Would you like to visit the LeoCAD website now?\n";
 
 							if (AfxMessageBox(str, MB_YESNO | MB_ICONQUESTION) == IDYES)
 							{
-								ShellExecute(::GetDesktopWindow(), _T("open"), _T("http://www.leocad.org/files/"), NULL, NULL, SW_NORMAL); 
+								ShellExecute(::GetDesktopWindow(), _T("open"), _T("http://www.leocad.org"), NULL, NULL, SW_NORMAL); 
 							}
 						}
 						else
@@ -125,12 +126,10 @@ static void CheckForUpdates(void* Data)
 	InternetCloseHandle(session);
 }
 
-CCADApp theApp;
-
 /////////////////////////////////////////////////////////////////////////////
 // CCADApp
 
-BEGIN_MESSAGE_MAP(CCADApp, CWinApp)
+BEGIN_MESSAGE_MAP(CCADApp, CWinAppEx)
 	//{{AFX_MSG_MAP(CCADApp)
 	ON_COMMAND(ID_HELP_CHECKFORUPDATES, OnHelpUpdates)
 	ON_COMMAND(ID_HELP_LEOCADHOMEPAGE, OnHelpHomePage)
@@ -138,12 +137,24 @@ BEGIN_MESSAGE_MAP(CCADApp, CWinApp)
 	//}}AFX_MSG_MAP
 	// Standard print setup command
 	ON_UPDATE_COMMAND_UI(ID_FILE_MRU_FILE1, OnUpdateRecentFileMenu)
-	ON_COMMAND(ID_FILE_PRINT_SETUP, CWinApp::OnFilePrintSetup)
+	ON_COMMAND(ID_FILE_PRINT_SETUP, CWinAppEx::OnFilePrintSetup)
 END_MESSAGE_MAP()
+
+/////////////////////////////////////////////////////////////////////////////
+// CCADApp construction
 
 CCADApp::CCADApp()
 {
+	m_hMutex = NULL;
 }
+
+/////////////////////////////////////////////////////////////////////////////
+// The one and only CCADApp object
+
+CCADApp theApp;
+
+/////////////////////////////////////////////////////////////////////////////
+// CCADApp initialization
 
 BOOL CCADApp::InitInstance()
 {
@@ -161,12 +172,20 @@ BOOL CCADApp::InitInstance()
 #endif
 
 	SetRegistryKey(_T("BT Software"));
-//	LoadStdProfileSettings();  // Load standard INI file options (including MRU)
+	LoadStdProfileSettings();
+
+	InitContextMenuManager();
+	InitShellManager();
+	InitKeyboardManager();
+	InitTooltipManager();
+	CMFCToolTipInfo ttParams;
+	ttParams.m_bVislManagerTheme = TRUE;
+	theApp.GetTooltipManager()->SetTooltipParams(AFX_TOOLTIP_TYPE_ALL, RUNTIME_CLASS(CMFCToolTipCtrl), &ttParams);
 
 	InitKeyboardShortcuts();
 
 	char app[LC_MAXPATH], *ptr;
-	GetModuleFileName(NULL, app, LC_MAXPATH);
+	GetModuleFileName (NULL, app, LC_MAXPATH);
 	ptr = strrchr(app,'\\');
 	if (ptr)
 		*(++ptr) = 0;
@@ -194,39 +213,54 @@ BOOL CCADApp::InitInstance()
 	m_nCmdShow = SW_HIDE;
 	pDocTemplate->OpenDocumentFile(NULL);
 
+	CMainFrame* MainFrame = (CMainFrame*)AfxGetMainWnd();
+	MainFrame->UpdateMenuAccelerators();
+
 	// Show something in the piece preview window.
 	PieceInfo* Info = lcGetPiecesLibrary()->FindPieceInfo("3005");
 	if (!Info)
 		Info = lcGetPiecesLibrary()->GetPieceInfo(0);
 
 	if (Info)
-		g_App->m_PiecePreview->SetSelection(Info);
-
-/*
-	char out[_MAX_PATH];
-	GetTempPath(_MAX_PATH, out);
-	strcat(out, "~LC*.lcd");
-
-	WIN32_FIND_DATA fd;
-	HANDLE fh = FindFirstFile(out, &fd);
-	if (fh != INVALID_HANDLE_VALUE)
 	{
-		if (char *ptr = strrchr(out, '\\')) *(ptr+1) = 0;
-		strcat(out, fd.cFileName);
-		if (AfxMessageBox(_T("LeoCAD found a file that was being edited while the program exited unexpectdly. Do you want to load it ?"), MB_YESNO) == IDNO)
-		{
-			if (AfxMessageBox(_T("Delete file ?"), MB_YESNO) == IDYES)
-				DeleteFile(out);
-		}
-		else
-		{
-			cmdInfo.m_nShellCommand = CCommandLineInfo::FileOpen;
-			cmdInfo.m_strFileName = out;
-		}
+		lcGetActiveProject()->SetCurrentPiece(Info);
+		MainFrame->m_wndPiecesBar.m_wndPiecePreview.SetPieceInfo(Info);
+		MainFrame->m_wndPiecesBar.m_wndPiecePreview.PostMessage(WM_PAINT);
 	}
 
-//	if (cmdInfo.m_strFileName.IsEmpty())
+/*
+	m_hMutex = CreateMutex(NULL, FALSE, _T("LeoCAD_Mutex"));
+	if (GetLastError() == ERROR_ALREADY_EXISTS)
+	{
 //		ParseCommandLine(cmdInfo);
+	}
+	else
+	{
+		char out[_MAX_PATH];
+		GetTempPath (_MAX_PATH, out);
+		strcat (out, "~LC*.lcd");
+
+		WIN32_FIND_DATA fd;
+		HANDLE fh = FindFirstFile(out, &fd);
+		if (fh != INVALID_HANDLE_VALUE)
+		{
+			if (char *ptr = strrchr (out, '\\')) *(ptr+1) = 0;
+			strcat (out, fd.cFileName);
+			if (AfxMessageBox (_T("LeoCAD found a file that was being edited while the program exited unexpectdly. Do you want to load it ?"), MB_YESNO) == IDNO)
+			{
+				if (AfxMessageBox (_T("Delete file ?"), MB_YESNO) == IDYES)
+					DeleteFile (out);
+			}
+			else
+			{
+				cmdInfo.m_nShellCommand = CCommandLineInfo::FileOpen;
+				cmdInfo.m_strFileName = out;
+			}
+		}
+
+//		if (cmdInfo.m_strFileName.IsEmpty())
+//			ParseCommandLine(cmdInfo);
+	}
 */
 
 	// The one and only window has been initialized, so show and update it.
@@ -239,44 +273,24 @@ BOOL CCADApp::InitInstance()
 	lcGetActiveProject()->HandleNotify(LC_ACTIVATE, 1);
 	lcGetActiveProject()->UpdateInterface();
 
-	main_window->UpdateMRU();
+	main_window->UpdateMRU ();
 
 	// Enable drag/drop open
 	m_pMainWnd->DragAcceptFiles();
 
 	lcGetActiveProject()->UpdateAllViews();
 
-	int CheckUpdates = AfxGetApp()->GetProfileInt("Settings", "CheckUpdates", 1);
-	if (CheckUpdates)
-	{
-		struct tm When;
-		__time64_t Now, Next;
-
-		if (CheckUpdates == 2)
-			CheckUpdates = 7;
-
-		memset(&When, 0, sizeof(When));
-		CString LastCheck = GetProfileString("Settings", "LastUpdate", NULL);
-		sscanf(LastCheck, "%d %d %d", &When.tm_mday, &When.tm_mon, &When.tm_year);
-		When.tm_mday = When.tm_mday + CheckUpdates;
-		Next = _mktime64(&When);
-
-		_time64(&Now);
-
-		if (Next < Now)
-		{
-			When = *_localtime64(&Now);
-			_beginthread(CheckForUpdates, 0, NULL);
-			LastCheck.Format("%d %d %d", When.tm_mday, When.tm_mon, When.tm_year);
-			WriteProfileString("Settings", "LastUpdate", LastCheck);
-		}
-	}
+	if (AfxGetApp()->GetProfileInt("Settings", "CheckUpdates", 1))
+		_beginthread(CheckForUpdates, 0, NULL);
 
 	return TRUE;
 }
 
 int CCADApp::ExitInstance() 
 {
+	if (m_hMutex != NULL)
+		ReleaseMutex(m_hMutex);
+
 	delete main_window;
 	main_window = NULL;
 
@@ -290,8 +304,17 @@ int CCADApp::ExitInstance()
 		FreeConsole();
 #endif
 
-	return CWinApp::ExitInstance();
+	return CWinAppEx::ExitInstance();
 }
+
+void CCADApp::UpdateMRU(char names[4][MAX_PATH])
+{
+	for (int iMRU = 0; iMRU < m_pRecentFileList->m_nSize; iMRU++)
+		m_pRecentFileList->m_arrNames[iMRU] = names[iMRU];
+}
+
+/////////////////////////////////////////////////////////////////////////////
+// CCADApp commands
 
 void CCADApp::OnHelpUpdates() 
 {
@@ -305,5 +328,5 @@ void CCADApp::OnHelpHomePage()
 
 void CCADApp::OnHelpEmail() 
 {
-	ShellExecute(::GetDesktopWindow(), _T("open"), _T("mailto:leonardo@centroin.com.br?subject=LeoCAD"), NULL, NULL, SW_NORMAL);
+	ShellExecute(::GetDesktopWindow(), _T("open"), _T("mailto:leozide@gmail.com?subject=LeoCAD"), NULL, NULL, SW_NORMAL);
 }
